Fix panic when iteration continues after error.
The point field is always incremented after every call to next. However
if the option was invalid, the error was returned to the user without
checking that the parser was in a valid state. This could lead to
a panic on the next call when index would then be out of bounds.

This change also checks that point >= the length on the error path,
incrementing it if necessary.